// src/app/features/top-news/models/news-item.model.ts
import { MultiMedia } from './multi-media.model';

export interface NewsItem {
  section: string;
  subsection: string;
  title: string;
  abstract: string;
  url: string;
  uri: string;
  byline: string;
  item_type: string;
  updated_date: string;
  created_date: string;
  published_date: string;
  material_type_facet: string;
  kicker: string;
  des_facet: string[];
  org_facet: string[];
  per_facet: string[];
  geo_facet: string[];
  multimedia: MultiMedia[];
  short_url: string;
}